requirements.txt
absl-py==0.6.1
astor==0.7.1
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
gast==0.2.0
grpcio==1.17.1
h5py==2.9.0
idna==2.8
isort==4.3.4
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
numpy==1.15.4
Pillow==5.3.0
protobuf==3.6.1
pylint==2.2.2
PyYAML==3.13
requests==2.21.0
scipy==1.2.0
six==1.12.0
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
typed-ast==1.1.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
